+++ inc/Helpers/StringsImplementation.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Airfleet\Framework\Helpers;
+
+use Jawira\CaseConverter\Convert;
+
+class StringsImplementation {
+	public function convert( string $source ): Convert {
+		return new Convert( $source );
+	}
+}